packages/testbench-ls/pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"


[project]
name = "testbench-ls"
authors = [
    {name = "imbus AG", email = "support@imbus.de"},
]
description = ""
requires-python = ">= 3.9"
readme = "README.md"
license = {text = "Apache 2.0 License"}
dynamic = ["version"]
dependencies = [
  "click >= 8.0",
  "robotframework >= 5.0",
  "pygls >= 1.3",
  "requests",
  "testbench-cli-reporter >= 2.0.0a4",
  "markdownify"
]

[project.optional-dependencies]
dev = [
  "build",
  "check-manifest",
  "mypy",
  "pylint",
  "pytest",
  "pytest-cov",
  "pytest-spec",
  "robotframework-tidy",
  "ruff",
  "twine",
]

[project.scripts]
testbench-ls = "testbench_ls.cli:testbench_ls_cli"

# [project.urls]
# Repository = ""

[tool.setuptools]
packages = ["testbench_ls", "testbench_ls.testbench_keysync"]

[tool.setuptools.dynamic]
version = {attr = "testbench_ls.__version__"}


[tool.check-manifest]
ignore = [
  "config.json",
  "*.bat",
  "*.sh",
  "*.zip",
  "tests/**"
]


[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true
no_implicit_optional = true
strict_optional = true
warn_return_any = true
warn_no_return = true
warn_unreachable = true
pretty = true


[tool.ruff]
target-version = "py39"
line-length = 100
lint.select = [
  "A",
  "B",
  "C",
  "DTZ",
  "E",
  "F",
  "I",
  "ICN",
  "INP",
  "ISC",
  "N",
  "PIE",
  "PL",
  "PT",
  "PTH",
  "PYI",
  "RET",
  "RSE",
  "RUF",
  "SIM",
  "T20",
  "UP",
  "W"
]

lint.ignore = [
  "ISC001"
]

exclude = [
  "__pycache__",
]


[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra --spec"
testpaths = [
    "tests",
]
spec_header_format = "{test_case}"


[tool.pylint.MASTER]
ignore = "tests"

[tool.pylint.BASIC]
argument-naming-style = 'snake_case'
attr-naming-style = 'snake_case'
class-naming-style = 'PascalCase'
const-naming-style = 'UPPER_CASE'
function-naming-style = 'snake_case'
method-naming-style = 'snake_case'
module-naming-style = 'snake_case'
variable-naming-style = 'snake_case'

[tool.pylint.DESIGN]
max-statements = 50

[tool.pylint."MESSAGES CONTROL"]
disable = '''
    logging-fstring-interpolation,
    fixme,
    import-error,
    missing-module-docstring,
    missing-class-docstring,
    missing-function-docstring,
    too-many-instance-attributes,
    too-few-public-methods,
'''

[tool.pylint.REFACTORING]
max-nested-blocks = 3
